requirements.txt
south
django-debug-toolbar
